카페리스트응답클래스 및 검색 dto, service수정, User클래스 수정,review클래스 수정

+++ src/main/java/com/cafe/cafeMood/tag/dto/TopTagResponse.java
@@ -3,6 +3,6 @@ package com.cafe.cafeMood.tag.dto;
 public record TopTagResponse(
         Long tagId,
         String tagName,
-        long score
+        double score
 ) {
 }
